异常规范
