--- exception/__init__.py
@@ -1,14 +1,0 @@
-# -*- coding:utf-8 -*-
-# @Time     : 2018-12-31 22:07
-# @Author   : 胡远
-# @Github   : https://github.com/QuixoteHY
-# @Email    : 1290482442@qq.com
-# @Describe : 自定义异常类
-
-
-class NoCallbackError(Exception):
-    pass
-
-
-class NoRequestInQueue(Exception):
-    pass
