--- exception/download_handlers.py
@@ -1,14 +1,0 @@
-# -*- coding:utf-8 -*-
-# @Time     : 2019-01-15 22:22
-# @Author   : 胡远
-# @Github   : https://github.com/QuixoteHY
-# @Email    : 1290482442@qq.com
-# @Describe :
-
-
-class NotSupportedURLScheme(Exception):
-    pass
-
-
-class NotConfigured(Exception):
-    pass
