--- exception/error.py
@@ -1,14 +1,0 @@
-# -*- coding:utf-8 -*-
-# @Time     : 2019-01-21 21:41
-# @Author   : 胡远
-# @Github   : https://github.com/QuixoteHY
-# @Email    : 1290482442@qq.com
-# @Describe :
-
-
-class ErrorNeedExit(Exception):
-    pass
-
-
-class ErrorSettings(ErrorNeedExit):
-    pass
